requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.7.4
requests>=2.32.2
python-dotenv>=1.0.0
langchain-openai>=0.2.0
langchain-core>=0.3.72
langgraph>=0.2.0
rich>=13.7.0
langchain-community>=0.3.0
faiss-cpu>=1.7.4
pandas>=2.0.0

